fix: created BillingDetails component to make it reusable
