+++ src/components/Billing/BillingDetails.vue
@@ -1,0 +1,58 @@
+<template>
+  <div class="flex flex-col gap-5">
+    <FormControl
+      v-model="billingInformation.billing_name"
+      type="text"
+      name="billing_name"
+      label="Billing Name"
+      :required="true"
+    />
+    <AddressForm
+      ref="addressFormRef"
+      v-model="billingInformation"
+      @success="() => emit('success')"
+    />
+  </div>
+</template>
+<script setup>
+import AddressForm from './AddressForm.vue'
+import FormControl from '../FormControl.vue'
+import { createResource } from '../../resources/index.js'
+import { reactive, ref, inject } from 'vue'
+
+const emit = defineEmits(['success'])
+const { baseAPIPath } = inject('billing')
+
+const addressFormRef = ref(null)
+
+const billingInformation = reactive({
+  billing_name: '',
+  address: '',
+  city: '',
+  state: '',
+  postal_code: '',
+  country: '',
+  gstin: '',
+})
+
+createResource({
+  url: `${baseAPIPath}.saas_api`,
+  params: { method: 'billing.get_information' },
+  auto: true,
+  onSuccess: (data) => {
+    Object.assign(billingInformation, {
+      address: data.address_line1,
+      city: data.city,
+      state: data.state,
+      postal_code: data.pincode,
+      country: data.country,
+      gstin: data.gstin == 'Not Applicable' ? '' : data.gstin,
+      billing_name: data.billing_name,
+    })
+  },
+})
+
+defineExpose({
+  getUpdateResource: () => addressFormRef.value.updateBillingInformation,
+})
+</script>
